refactor: extract state snapshot CRUD from message-handlers
Moved browser state snapshot persistence (save/load/list/delete) into
dedicated state-snapshots.ts module. message-handlers.ts now focuses
solely on chrome.runtime.onMessage routing.

Re-exports preserved for backward compatibility. Cleaned up unused
imports (BrowserStateSnapshot, setLocal) from message-handlers.ts.
