+++ src/background/message-handlers.ts
@@ -13,7 +13,6 @@ import type {
   BackgroundMessage,
   DrawModeCompletedMessage,
   ChromeMessageSender,
-  BrowserStateSnapshot,
   ConnectionStatus,
   ContextWarning,
   CircuitBreakerState,
@@ -27,7 +26,7 @@ import { SettingName, StorageKey, DEFAULT_SERVER_URL } from '../lib/constants.js
 import { pushChatMessage } from './push-handler.js'
 import { errorMessage } from '../lib/error-utils.js'
 import { postDaemonJSON } from '../lib/daemon-http.js'
-import { getLocal, getLocals, setLocal } from '../lib/storage-utils.js'
+import { getLocal, getLocals } from '../lib/storage-utils.js'
 
 // =============================================================================
 // TYPE DEFINITIONS
@@ -577,67 +576,6 @@ function handleSetServerUrl(url: string, sendResponse: SendResponse, deps: Messa
 }
 
 // =============================================================================
-// STATE SNAPSHOT STORAGE
-// =============================================================================
-
-const SNAPSHOT_KEY = 'gasoline_state_snapshots'
-
-interface StoredStateSnapshot extends BrowserStateSnapshot {
-  name: string
-  size_bytes: number
-}
-
-interface StateSnapshotStorage {
-  [name: string]: StoredStateSnapshot
-}
-
-/**
- * Save a state snapshot to persistent storage
- */
-export async function saveStateSnapshot(
-  name: string,
-  state: BrowserStateSnapshot
-): Promise<{ success: boolean; snapshot_name: string; size_bytes: number }> {
-  const existing = (await getLocal(SNAPSHOT_KEY)) as StateSnapshotStorage | undefined
-  const snapshots: StateSnapshotStorage = existing || {}
-  const sizeBytes = JSON.stringify(state).length // nosemgrep: no-stringify-keys
-  snapshots[name] = { ...state, name, size_bytes: sizeBytes }
-  await setLocal(SNAPSHOT_KEY, snapshots)
-  return { success: true, snapshot_name: name, size_bytes: sizeBytes }
-}
-
-/**
- * Load a state snapshot from persistent storage
- */
-export async function loadStateSnapshot(name: string): Promise<StoredStateSnapshot | null> {
-  const existing = (await getLocal(SNAPSHOT_KEY)) as StateSnapshotStorage | undefined
-  const snapshots: StateSnapshotStorage = existing || {}
-  return snapshots[name] || null
-}
-
-/**
- * List all state snapshots with metadata
- */
-export async function listStateSnapshots(): Promise<
-  Array<{ name: string; url: string; timestamp: number; size_bytes: number }>
-> {
-  const existing = (await getLocal(SNAPSHOT_KEY)) as StateSnapshotStorage | undefined
-  const snapshots: StateSnapshotStorage = existing || {}
-  return Object.values(snapshots).map((s) => ({
-    name: s.name,
-    url: s.url,
-    timestamp: s.timestamp,
-    size_bytes: s.size_bytes
-  }))
-}
-
-/**
- * Delete a state snapshot from persistent storage
- */
-export async function deleteStateSnapshot(name: string): Promise<{ success: boolean; deleted: string }> {
-  const existing = (await getLocal(SNAPSHOT_KEY)) as StateSnapshotStorage | undefined
-  const snapshots: StateSnapshotStorage = existing || {}
-  delete snapshots[name]
-  await setLocal(SNAPSHOT_KEY, snapshots)
-  return { success: true, deleted: name }
-}
+// STATE SNAPSHOT STORAGE — delegated to state-snapshots.ts
+// Re-export for backward compatibility
+export { saveStateSnapshot, loadStateSnapshot, listStateSnapshots, deleteStateSnapshot } from './state-snapshots.js'
